Rename robot module
